# add bench for depth performances