related to #71 

on my local machine, took 7m24s:
```
cargo bench --bench simple_deep
   Compiling stretch v0.3.2 (/Users/francois/Dev/spaceward-ho/stretch)
warning: unused `std::result::Result` that must be used
  --> benches/simple_deep.rs:48:13
   |
48 |             stretch.mark_dirty(root);
   |             ^^^^^^^^^^^^^^^^^^^^^^^^^
   |
   = note: `#[warn(unused_must_use)]` on by default
   = note: this `Result` may be an `Err` variant, which should be handled

warning: 1 warning emitted

    Finished bench [optimized] target(s) in 2.53s
     Running target/release/deps/simple_deep-c97ff526f7fd7766
simple deep hierarchy - build
                        time:   [9.3641 us 9.4032 us 9.4477 us]
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe

Benchmarking simple deep hierarchy - single: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 210.9s, or reduce sample count to 10.
simple deep hierarchy - single
                        time:   [2.0722 s 2.0818 s 2.0922 s]
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe

Benchmarking simple deep hierarchy - relayout: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 208.3s, or reduce sample count to 10.
simple deep hierarchy - relayout
                        time:   [2.0639 s 2.0697 s 2.0756 s]
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild
```